requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-marshmallow 
flask-cors
apifairy
python-dotenv
marshmallow-sqlalchemy